# packages/backend/swagger.json
{"openapi":"3.0.0","paths":{"/users/{userId}":{"get":{"operationId":"UserController_get","summary":"Get a user","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["UserService"]},"patch":{"operationId":"UserController_update","summary":"Update a user","parameters":[{"name":":userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequestBody"}}}},"responses":{"200":{"description":""}},"tags":["UserService"]},"delete":{"operationId":"UserController_deactivate","summary":"Delete a user","parameters":[{"name":":userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["UserService"]}},"/users":{"get":{"operationId":"UserController_list","summary":"List users","parameters":[{"name":"page","required":false,"in":"query","description":"ページネーション: ページ番号","schema":{"default":1,"type":"number"}},{"name":"perPage","required":false,"in":"query","description":"ページネーション: 1ページのデータ数","schema":{"default":10,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["UserService"]},"post":{"operationId":"UserController_create","summary":"Create a new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequestBody"}}}},"responses":{"201":{"description":""}},"tags":["UserService"]}}},"info":{"title":"Account Viewer","description":"Account Viewer API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserRequestBody":{"type":"object","properties":{"displayName":{"type":"string","description":"display name","example":"Yamada Tarou"},"email":{"type":"string","description":"email address","example":"taro@sample.com"}},"required":["displayName","email"]},"UpdateUserRequestBody":{"type":"object","properties":{"displayName":{"type":"string","description":"display name","example":"Yamada Tarou"},"email":{"type":"string","description":"email address","example":"taro@sample.com"}}}}}}